Fix issue with product generation
This patch addresses an issue that would manifest when large numbers
were given for the STATIC_PRODUCT_COUNT. The script issued curls to the
WooCommerce Product Generator plugin with the `max` parameter set to
one-fifth of the requested total. However, the default PHP setup will
timeout a request after 30 seconds, such that the number of products
generated was far fewer than supposed.

The patch addresses this my querying the database for the actual number
of products and continuing to issue curl requests until the parameter is
satisfied. Additionally, these requests are now performed in parallel by
curl processes -- this better tests the database under load and should
reduce the total product generation time.