Reduce number of generated users
This patch reduces the number of users generated as part of the base
Wordpress installation and the WooCommerce installation. These values
have not shown any impact on the benchmark's results and can take a long
time to complete. Reducing these numbers improves the benchmark's
runtime without affecting the results.

The patch also removes the `sleep`s in realtime product and comment
generation, and adds a POST request to the WordPress login page to the
list of URLs to Siege.