Fix issue with capture_metrics redirection
This patch fixes an issue that would occur when executing the benchmark
with fully remote hosts (ie, main.sh is not running on a server). The
redirect in `capture_metrics` was outside the remote command, such that
bash tried to write the file on the local machine rather than the remote
server in question.